# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Job Vacancy API
  version: 0.6.0
  description: |
    Shared contract for the web (React) and mobile (Flutter) clients.

    **Realtime.** POST /ws/ticket, then open `wss://…/api/v1/ws?ticket=…`. Server events:
    `ready`, `message.new`, `message.read`, `message.deleted`, `typing`, `presence`,
    `presence.snapshot`, `notification.new`, `pong`, `error` — each `{"type", "data"}`.
    Client frames: `{"type":"ping"}`, `{"type":"typing","conversation_id"}`,
    `{"type":"read","conversation_id","message_id"}`, `{"type":"watch_presence","user_ids":[…]}`.
    Events are not buffered: after reconnecting, fetch `/conversations/{id}/messages?after=<last id>`.

    **Uploads.** POST /files → multipart POST to `upload.url` with every `upload.fields`
    entry followed by the `file` field → POST /files/{id}/complete → use the file id.

    **Envelope.** Success: `{"data": ...}` (lists add `"meta"`). Failure:
    `{"error": {"code", "message", "fields"?, "retry_after"?}}`. `code` is stable; clients
    translate it into uz / uz-Cyrl / ru / en. `fields` maps a JSON field to the failing
    rule (`required`, `email`, `min=8`, `password`, `oneof=...`).

    **Deadlines.** Every request runs under a server deadline (8 s; PDF export 25 s;
    the WebSocket is exempt) and each database statement under 5 s. When one is exceeded
    the API answers `503` with code `timeout` and `Retry-After: 1`; the call is safe to
...
    limits of their own.

...

...
    it rendered.

...

...
servers:
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    ClientType:
      name: X-Client-Type
      in: header
      schema: { type: string, enum: [web, android, ios], default: web }
  schemas:
    Locale: { type: string, enum: [uz, uz-Cyrl, ru, en] }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [invalid_credentials] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            retry_after: { type: integer, description: seconds }
    User:
      type: object
      required: [id, email, email_verified, phone, phone_verified, full_name, avatar_url, role, locale, has_password, google_linked, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        email_verified: { type: boolean }
        phone: { type: [string, "null"], examples: ["+998901234567"] }
        phone_verified: { type: boolean }
        full_name: { type: string }
        avatar_url: { type: [string, "null"] }
        role: { type: string, enum: [seeker, employer, admin] }
        locale: { $ref: "#/components/schemas/Locale" }
        has_password: { type: boolean }
        google_linked: { type: boolean }
        created_at: { type: string, format: date-time }
        hide_online: { type: boolean, description: "online status hidden from everyone (TZ SEC-05)" }
        consent_version: { type: [string, "null"], description: "privacy policy version accepted (TZ FN-08); null for accounts created before consent was recorded" }
        consent_at: { type: [string, "null"], format: date-time }
        consent_outdated: { type: boolean, description: "true: show the current policy and PUT /me/consent" }
    Auth:
      type: object
      properties:
        data:
          type: object
          properties:
            user: { $ref: "#/components/schemas/User" }
            access_token: { type: string }
            access_token_expires_at: { type: string, format: date-time }
            refresh_token: { type: string, description: mobile clients only }
            refresh_token_expires_at: { type: string, format: date-time, description: mobile clients only }
            is_new_user: { type: boolean }
    UserResponse:
      type: object
      properties: { data: { $ref: "#/components/schemas/User" } }
...
    Code:
      type: object
      required: [code]
      properties: { code: { type: string, pattern: "^[0-9]{6}$" } }
...
  responses:
...
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Auth:
      description: Signed in
      content: { application/json: { schema: { $ref: "#/components/schemas/Auth" } } }
    User:
      description: Current user
      content: { application/json: { schema: { $ref: "#/components/schemas/UserResponse" } } }
    NoContent: { description: Done }
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Sign up with email and password (sends a verification code by email)
      parameters: [{ $ref: "#/components/parameters/ClientType" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name, role]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 72, description: at least one letter and one digit }
                full_name: { type: string, minLength: 2, maxLength: 100 }
                role: { type: string, enum: [seeker, employer] }
                locale: { $ref: "#/components/schemas/Locale" }
                consent: { type: boolean, description: "consent to the processing of personal data (TZ FN-08); must be true, else 400 consent_required" }
                consent_version: { type: string, maxLength: 32, description: "policy version shown; omitted = current; an older one is 409 consent_outdated" }
      responses:
        "201": { $ref: "#/components/responses/Auth" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in with email and password
      parameters: [{ $ref: "#/components/parameters/ClientType" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Auth" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/google:
    post:
      tags: [auth]
      summary: Sign in with a Google ID token (creates the account on first use)
      parameters: [{ $ref: "#/components/parameters/ClientType" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token: { type: string }
                role: { type: string, enum: [seeker, employer], description: used only for new accounts }
                locale: { $ref: "#/components/schemas/Locale" }
                consent: { type: boolean, description: "required (true) when this sign-in creates the account, else 400 consent_required" }
                consent_version: { type: string, maxLength: 32 }
      responses:
        "200": { $ref: "#/components/responses/Auth" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token and get a new access token
      description: |
        `refresh_race` (401) means a parallel request already rotated the token: retry once.
        Reusing an old refresh token later revokes the whole session.
      parameters: [{ $ref: "#/components/parameters/ClientType" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { refresh_token: { type: string, description: mobile only } }
      responses:
        "200": { $ref: "#/components/responses/Auth" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the session that owns the given refresh token
      parameters: [{ $ref: "#/components/parameters/ClientType" }]
      responses: { "204": { $ref: "#/components/responses/NoContent" } }
  /auth/logout-all:
    post:
      tags: [auth]
      security: [{ bearer: [] }]
      summary: Revoke every session of the current user
      responses: { "204": { $ref: "#/components/responses/NoContent" } }
  /auth/password/forgot:
    post:
      tags: [auth]
      summary: Email a password reset code (always 204, never reveals if the email exists)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email], properties: { email: { type: string, format: email } } }
      responses: { "204": { $ref: "#/components/responses/NoContent" } }
  /auth/password/reset:
    post:
      tags: [auth]
      summary: Set a new password with the emailed code; signs out all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, password]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{6}$" }
                password: { type: string, minLength: 8, maxLength: 72 }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/Error" }
  /auth/email/send-code:
    post:
      tags: [auth]
      security: [{ bearer: [] }]
      summary: (Re)send the email verification code (60 s cooldown)
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/email/verify:
    post:
      tags: [auth]
      security: [{ bearer: [] }]
      summary: Confirm the email with the 6-digit code
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/Code" } } } }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/Error" }
  /auth/phone/send-code:
    post:
      tags: [auth]
      security: [{ bearer: [] }]
      summary: Send a phone verification code through Telegram ("Verification Codes" chat)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [phone], properties: { phone: { type: string, examples: ["+998 90 123 45 67"] } } }
      responses:
        "202":
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      phone: { type: string, description: normalized E.164 }
                      channel: { type: string, const: telegram }
                      expires_in: { type: integer, description: seconds }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/phone/verify:
    post:
      tags: [auth]
      security: [{ bearer: [] }]
      summary: Confirm the phone number with the code from Telegram
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/Code" } } } }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/Error" }
...